refactor filter to its own component
